+++ mail_to/models/mail_message.py
@@ -13,5 +13,16 @@ class MailMessage(models.Model):
             if i['channel_ids']:
                 i['channel_names'] = self.env['mail.channel'].browse(i['channel_ids']).mapped(
                     lambda r: [r.id, '#' + r.display_name])
+            else:
+                i['channel_names'] = []
+
+            partner_ids = set(i['needaction_partner_ids'])
+            partner_ids.update(set(map(lambda x: x[0], i['partner_ids'])))
+            partner_ids.update(set(map(lambda x: x[0], i['customer_email_data'])))
+            if partner_ids:
+                i['partner_names'] = self.env['res.partner'].browse(partner_ids).mapped(
+                    lambda r: [r.id, r.name])
+            else:
+                i['partner_names'] = []
 
         return messages_values