+++ mail_to/doc/changelog.rst
@@ -1,8 +1,7 @@
 `1.1.0`
 -------
 
-- **New**: Channels are displayed in recipients
-
+- **New**: channels are displayed in recipients
 
 `1.0.1`
 -------